call data client
