+++ packages/core/src/client.ts
@@ -1,5 +1,4 @@
-import { decodeEventLog, encodeEventTopics, encodeFunctionData } from "viem";
-import { HatsSubgraphClient } from "@hatsprotocol/sdk-v1-subgraph";
+import { decodeEventLog, encodeEventTopics } from "viem";
 
 import { HATS_ABI } from "./abi/Hats";
 import {
@@ -11,8 +10,8 @@ import {
   MaxHatsInLevelReached,
 } from "./errors";
 import { HATS_V1, MAX_LEVEL_HATS, MAX_LEVELS, ZERO_ADDRESS } from "./constants";
-import { treeIdDecimalToHex, hatIdHexToDecimal } from "./utils";
 import { getError } from "./validations";
+import { HatsCallDataClient } from "./calldata";
 import type { PublicClient, WalletClient, Account, Address, Hex } from "viem";
 import type {
   CreateHatResult,
@@ -39,11 +38,8 @@ import type {
   MultiCallResult,
 } from "./types";
 
-export class HatsClient {
-  readonly chainId: number;
-  private readonly _publicClient: PublicClient;
+export class HatsClient extends HatsCallDataClient {
   private readonly _walletClient: WalletClient | undefined;
-  private readonly _graphqlClient: HatsSubgraphClient | undefined;
 
   /**
    * Initialize a HatsClient.
@@ -68,6 +64,7 @@ export class HatsClient {
     publicClient: PublicClient;
     walletClient?: WalletClient;
   }) {
+    super({ chainId, publicClient });
     if (publicClient === undefined) {
       throw new MissingPublicClientError("Public client is required");
     }
@@ -84,9 +81,6 @@ export class HatsClient {
       );
     }
 
-    this.chainId = chainId;
-    this._graphqlClient = new HatsSubgraphClient();
-    this._publicClient = publicClient;
     this._walletClient = walletClient;
   }
 
@@ -469,14 +463,11 @@ export class HatsClient {
    * @param account - A Viem account.
    * @param target - Tophat's wearer address.
    * @param details - Tophat's details field.
-   * @param imageURIi - Optional tophat's image URI.
+   * @param imageURI - Optional tophat's image URI.
    * @returns An object containing the status of the call, the transactions hash and the created tophat ID.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpeced reason.
+   * Thrown if no wallet client was provided in the hats client initialization.
    */
   async mintTopHat({
     account,
@@ -495,35 +486,31 @@ export class HatsClient {
       );
     }
 
-    try {
-      const hash = await this._walletClient.writeContract({
-        address: HATS_V1,
-        abi: HATS_ABI,
-        functionName: "mintTopHat",
-        args: [target, details, imageURI === undefined ? "" : imageURI],
-        account,
-        chain: this._walletClient.chain,
-      });
+    const hash = await this._walletClient.writeContract({
+      address: HATS_V1,
+      abi: HATS_ABI,
+      functionName: "mintTopHat",
+      args: [target, details, imageURI === undefined ? "" : imageURI],
+      account,
+      chain: this._walletClient.chain,
+    });
 
-      const receipt = await this._publicClient.waitForTransactionReceipt({
-        hash,
-      });
+    const receipt = await this._publicClient.waitForTransactionReceipt({
+      hash,
+    });
 
-      const event = decodeEventLog({
-        abi: HATS_ABI,
-        eventName: "HatCreated",
-        data: receipt.logs[0].data,
-        topics: receipt.logs[0].topics,
-      });
+    const event = decodeEventLog({
+      abi: HATS_ABI,
+      eventName: "HatCreated",
+      data: receipt.logs[0].data,
+      topics: receipt.logs[0].topics,
+    });
 
-      return {
-        status: receipt.status,
-        transactionHash: receipt.transactionHash,
-        hatId: event.args.id,
-      };
-    } catch (err) {
-      throw new TransactionRevertedError("Transaction reverted");
-    }
+    return {
+      status: receipt.status,
+      transactionHash: receipt.transactionHash,
+      hatId: event.args.id,
+    };
   }
 
   /**
@@ -532,7 +519,7 @@ export class HatsClient {
    * @param account - A Viem account.
    * @param admin - Hat's admin ID.
    * @param details - Hat's details field.
-   * @param maxSupply - Hat's maximum amount of possible wearers.
+   * @param maxSupply - Hat's maximum amount of wearers.
    * @param eligibility - Hat's eligibility address (zero address is not valid).
    * @param toggle - Hat's toggle address (zero address is not valid).
    * @param mutable - True if the hat should be mutable, false otherwise.
@@ -540,23 +527,16 @@ export class HatsClient {
    * @returns An object containing the status of the call, the transaction hash and the created hat ID.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
-   * @throws ZeroEligibilityError
-   * Throws if provided the zero address as an eligibility.
-   *
-   * @throws ZeroToggleError
-   * Throws if provided the zero address as a toggle.
+   * @throws ZeroAddressError
+   * Thrown if provided the zero address as an eligibility or toggle.
    *
    * @throws InvalidAdminError
-   * Throws if the provided admin ID is not valid.
+   * Thrown if the provided admin ID is not valid.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of the hat that will be created.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
-   *
+   * Thrown if the calling account is not an admin of the hat that will be created.
    */
   async createHat({
     account,
@@ -629,7 +609,7 @@ export class HatsClient {
    * @param account - A Viem account.
    * @param admins - The hats admin IDs.
    * @param details - The hats details fields.
-   * @param maxSupplies - The hats maximum amounts of possible wearers.
+   * @param maxSupplies - The hats maximum amounts of wearers.
    * @param eligibilityModules - The hats eligibility addresses (zero address is not valid).
    * @param toggleModules - The hats toggle addresses (zero address is not valid).
    * @param mutables - True if the hat should be mutable, false otherwise.
@@ -637,26 +617,19 @@ export class HatsClient {
    * @returns An object containing the status of the call, the transaction hash and the created hat IDs.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
-   * @throws ZeroEligibilityError
-   * Throws if provided the zero address as an eligibility.
-   *
-   * @throws ZeroToggleError
-   * Throws if provided the zero address as a toggle.
+   * @throws ZeroAddressError
+   * Thrown if provided the zero address as an eligibility or toggle.
    *
    * @throws InvalidAdminError
-   * Throws if the provided admin ID is not valid.
+   * Thrown if the provided admin ID is not valid.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of the hat that will be created.
+   * Thrown if the calling account is not an admin of the hat that will be created.
    *
    * @throws BatchParamsError
-   * Throws if there is a length mismatch between the provided hats properties.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
-   *
+   * Thrown if there is a length mismatch between the provided hats properties.
    */
   async batchCreateHats({
     account,
@@ -738,28 +711,25 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of hat.
+   * Thrown if the calling account is not an admin of hat.
    *
    * @throws HatNotExistError
-   * Throws if the hat does not exist.
+   * Thrown if the hat does not exist.
    *
    * @throws AllHatsWornError
-   * Throws if all the hats of the provided hat ID are currently worn.
+   * Thrown if all the hats of the provided hat ID are currently worn.
    *
    * @throws NotEligibleError
-   * Throws if the new wearer is not eligible for the hat.
+   * Thrown if the new wearer is not eligible for the hat.
    *
    * @throws NotActiveError
-   * Throws if the hat is not active.
+   * Thrown if the hat is not active.
    *
    * @throws AlreadyWearingError
-   * Throws if the new wearer is already wearing the hat.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if the new wearer is already wearing the hat.
    *
    */
   async mintHat({
@@ -810,32 +780,28 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of hat.
+   * Thrown if the calling account is not an admin of hat.
    *
    * @throws HatNotExistError
-   * Throws if the hat does not exist.
+   * Thrown if the hat does not exist.
    *
    * @throws AllHatsWornError
-   * Throws if all the hats of the provided hat ID are currently worn.
+   * Thrown if all the hats of the provided hat ID are currently worn.
    *
    * @throws NotEligibleError
-   * Throws if the new wearer is not eligible for the hat.
+   * Thrown if the new wearer is not eligible for the hat.
    *
    * @throws NotActiveError
-   * Throws if the hat is not active.
+   * Thrown if the hat is not active.
    *
    * @throws AlreadyWearingError
-   * Throws if the new wearer is already wearing the hat.
+   * Thrown if the new wearer is already wearing the hat.
    *
-   * * @throws BatchParamsError
-   * Throws if there is a length mismatch between the provided properties.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
-   *
+   * @throws BatchParamsError
+   * Thrown if there is a length mismatch between the provided properties.
    */
   async batchMintHats({
     account,
@@ -885,13 +851,10 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotToggleError
-   * Throws if the calling account is not the toggle of the hat.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if the calling account is not the toggle of the hat.
    */
   async setHatStatus({
     account,
@@ -941,10 +904,7 @@ export class HatsClient {
    * toggled and the new status in case the status was toggled.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if no wallet client was provided in the hats client initialization.
    */
   async checkHatStatus({
     account,
@@ -1010,13 +970,10 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotEligibilityError
-   * Throws if the calling account is not the eligibility of the hat.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if the calling account is not the eligibility of the hat.
    */
   async setHatWearerStatus({
     account,
@@ -1072,10 +1029,7 @@ export class HatsClient {
    * was updated, indicator whether the wearer's hat was burned and if standing has changed then the new standing.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if no wallet client was provided in the hats client initialization.
    */
   async checkHatWearerStatus({
     account,
@@ -1181,10 +1135,7 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if no wallet client was provided in the hats client initialization.
    */
   async renounceHat({
     account,
@@ -1233,28 +1184,25 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws ImmutableHatError
-   * Throws if the hat is immutable. Immutable hats cannot be transfered.
+   * Thrown if the hat is immutable. Immutable hats cannot be transfered.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of the hat.
+   * Thrown if the calling account is not an admin of the hat.
    *
    * @throws NotEligibleError
-   * Throws if the new wearer is not eligible for the hat.
+   * Thrown if the new wearer is not eligible for the hat.
    *
    * @throws NotActiveError
-   * THrows if the hat is not active.
+   * Thrown if the hat is not active.
    *
    * @throws NotWearerError
-   * Throws if the provided current wearer is not wearing the hat.
+   * Thrown if the provided current wearer is not wearing the hat.
    *
    * @throws AlreadyWearingError
-   * Throws if the new wearer is already wearing the hat.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if the new wearer is already wearing the hat.
    */
   async transferHat({
     account,
@@ -1305,16 +1253,13 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of the hat.
+   * Thrown if the calling account is not an admin of the hat.
    *
    * @throws ImmutableHatError
-   * Throws if the hat is immutable. Immutable hats cannot be edited.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if the hat is immutable. Immutable hats cannot be edited.
    */
   async makeHatImmutable({
     account,
@@ -1362,19 +1307,16 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of the hat.
+   * Thrown if the calling account is not an admin of the hat.
    *
    * @throws ImmutableHatError
-   * Throws if the hat is immutable.
+   * Thrown if the hat is immutable.
    *
    * @throws StringTooLongError
-   * Throws if the new details length is larger than 7000.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if the new details length is larger than 7000.
    */
   async changeHatDetails({
     account,
@@ -1424,19 +1366,16 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of the hat.
+   * Thrown if the calling account is not an admin of the hat.
    *
    * @throws ImmutableHatError
-   * Throws if the hat is immutable.
+   * Thrown if the hat is immutable.
    *
-   * @throws ZeroEligibilityError
-   * Throws if the new eligibilty is the zero address.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * @throws ZeroAddressError
+   * Thrown if the new eligibilty is the zero address.
    */
   async changeHatEligibility({
     account,
@@ -1486,19 +1425,16 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of the hat.
+   * Thrown if the calling account is not an admin of the hat.
    *
    * @throws ImmutableHatError
-   * Throws if the hat is immutable.
+   * Thrown if the hat is immutable.
    *
-   * @throws ZeroToggleError
-   * Throws if the new toggle is the zero address.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * @throws ZeroAddressError
+   * Thrown if the new toggle is the zero address.
    */
   async changeHatToggle({
     account,
@@ -1548,19 +1484,16 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of the hat.
+   * Thrown if the calling account is not an admin of the hat.
    *
    * @throws ImmutableHatError
-   * Throws if the hat is immutable.
+   * Thrown if the hat is immutable.
    *
    * @throws StringTooLongError
-   * Throws if the new image URI length is larger than 7000.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if the new image URI length is larger than 7000.
    */
   async changeHatImageURI({
     account,
@@ -1610,19 +1543,16 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of the hat.
+   * Thrown if the calling account is not an admin of the hat.
    *
    * @throws ImmutableHatError
-   * Throws if the hat is immutable.
+   * Thrown if the hat is immutable.
    *
    * @throws InvalidMaxSupplyError
-   * Throws if the new maximum supply is smaller the current amount of wearers.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if the new maximum supply is smaller the current amount of wearers.
    */
   async changeHatMaxSupply({
     account,
@@ -1672,13 +1602,10 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of the tophat.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if the calling account is not an admin of the tophat.
    */
   async requestLinkTopHatToTree({
     account,
@@ -1732,28 +1659,25 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NoLinkageRequestError
-   * Throws if the tophat has not requested the link.
+   * Thrown if the tophat has not requested the link.
    *
    * @throws NotAdminOrWearerError
-   * Throws if the calling account is not an admin or a wearer of the new admin hat.
+   * Thrown if the calling account is not an admin or a wearer of the new admin hat.
    *
    * @throws CircularLinkageError
-   * Throws if linking the trees creates a circular linkage.
+   * Thrown if linking the trees creates a circular linkage.
    *
    * @throws CrossLinkageError
-   * Throws if the new admin hat is in a different global tree than the current global
+   * Thrown if the new admin hat is in a different global tree than the current global
    * tree of the tophat that is being linked or if the calling account has no permission
    * to relink to the new destination within the same global tree.
    *
-   * * @throws StringTooLongErrorError
-   * Throws if a new details or new image URI were provided and either length is greater
+   * @throws StringTooLongErrorError
+   * Thrown if a new details or new image URI were provided and either length is greater
    * than 7000.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
    */
   async approveLinkTopHatToTree({
     account,
@@ -1818,16 +1742,13 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * @throws NotWearerError
-   * Throws if provided wearer is not the wearer of the tophat.
+   * Thrown if provided wearer is not the wearer of the tophat.
    *
    * @throws NotAdminError
-   * Throws if the calling account is not an admin of the tophat.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if the calling account is not an admin of the tophat.
    */
   async unlinkTopHatFromTree({
     account,
@@ -1881,28 +1802,25 @@ export class HatsClient {
    * @returns An object containing the status of the call and the transaction hash.
    *
    * @throws MissingWalletClientError
-   * Throws if no wallet client was provided in the hats client initialization.
+   * Thrown if no wallet client was provided in the hats client initialization.
    *
    * * @throws NotAdminError
-   * Throws if the calling account is not an admin of the tophat that is about to be relinked.
+   * Thrown if the calling account is not an admin of the tophat that is about to be relinked.
    *
    * @throws NotAdminOrWearerError
-   * Throws if the calling account is not an admin or a wearer of the new admin hat.
+   * Thrown if the calling account is not an admin or a wearer of the new admin hat.
    *
    * @throws CircularLinkageError
-   * Throws if linking the trees creates a circular linkage.
+   * Thrown if linking the trees creates a circular linkage.
    *
    * @throws CrossLinkageError
-   * Throws if the new admin hat is in a different global tree than the current global
+   * Thrown if the new admin hat is in a different global tree than the current global
    * tree of the tophat that is being linked or if the calling account has no permission
    * to relink to the new destination within the same global tree.
    *
    * * @throws StringTooLongErrorError
-   * Throws if a new details or new image URI were provided and either length is greater
+   * Thrown if a new details or new image URI were provided and either length is greater
    * than 7000.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
    */
   async relinkTopHatWithinTree({
     account,
@@ -1964,10 +1882,7 @@ export class HatsClient {
    * @returns An object containing newly created, minted or burned hats and hat/wearer status changes as a result of the multicall
    *
    * @throws MultiCallError
-   * Throws if the multicall simulation reverted.
-   *
-   * @throws TransactionRevertedError
-   * Throws if the transaction reverted for an unexpected reason.
+   * Thrown if the multicall simulation reverted.
    */
   async multicall({
     account,
@@ -2085,722 +2000,5 @@ export class HatsClient {
     } catch (err) {
       getError(err);
     }
-  }
-
-  /*//////////////////////////////////////////////////////////////
-                        Call Data functions
-      //////////////////////////////////////////////////////////////*/
-
-  /**
-   * Return the call data a multicall operation.
-   *
-   * @param calls - An array with the call data strings, for each function call.
-   * @returns An object containing the call data and the function name.
-   */
-  multicallCallData(calls: Hex[]): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "multicall",
-      args: [calls],
-    });
-
-    return { functionName: "multicall", callData };
-  }
-
-  /**
-   * Return the call data for a mintTopHat operation.
-   *
-   * @param target - Tophat's wearer address.
-   * @param details - Tophat's details field.
-   * @param imageURIi - Optional tophat's image URI.
-   * @returns An object containing the call data and the function name.
-   */
-  mintTopHatCallData({
-    target,
-    details,
-    imageURI,
-  }: {
-    target: Address;
-    details: string;
-    imageURI?: string;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "mintTopHat",
-      args: [target, details, imageURI === undefined ? "" : imageURI],
-    });
-
-    return { functionName: "mintTopHat", callData };
-  }
-
-  /**
-   * Return the call data for a createHat operation.
-   *
-   * @param admin - Hat's admin ID.
-   * @param details - Hat's details field.
-   * @param maxSupply - Hat's maximum amount of possible wearers.
-   * @param eligibility - Hat's eligibility address (zero address is not valid).
-   * @param toggle - Hat's toggle address (zero address is not valid).
-   * @param mutable - True if the hat should be mutable, false otherwise.
-   * @param imageURI - Optional hat's image URI.
-   * @returns An object containing the call data and the function name.
-   */
-  createHatCallData({
-    admin,
-    details,
-    maxSupply,
-    eligibility,
-    toggle,
-    mutable,
-    imageURI,
-  }: {
-    admin: bigint;
-    details: string;
-    maxSupply: number;
-    eligibility: Address;
-    toggle: Address;
-    mutable: boolean;
-    imageURI?: string;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "createHat",
-      args: [
-        admin,
-        details,
-        maxSupply,
-        eligibility,
-        toggle,
-        mutable,
-        imageURI === undefined ? "" : imageURI,
-      ],
-    });
-
-    return { functionName: "createHat", callData };
-  }
-
-  /**
-   * Return the call data for a transferHat operation.
-   *
-   * @param hatId - Hat ID to be transfered.
-   * @param from - Current wearer address.
-   * @param to - New wearer address.
-   * @returns An object containing the call data and the function name.
-   */
-  transferHatCallData({
-    hatId,
-    from,
-    to,
-  }: {
-    hatId: bigint;
-    from: Address;
-    to: Address;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "transferHat",
-      args: [hatId, from, to],
-    });
-
-    return { functionName: "transferHat", callData };
-  }
-
-  /**
-   * Return the call data for a mintHat operation.
-   *
-   * @param hatId - ID of the minted hat.
-   * @param wearer - Address of the new wearer.
-   * @returns An object containing the call data and the function name.
-   */
-  mintHatCallData({ hatId, wearer }: { hatId: bigint; wearer: Address }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "mintHat",
-      args: [hatId, wearer],
-    });
-
-    return { functionName: "mintHat", callData };
-  }
-
-  /**
-   * Return the call data for a batchCreateHats operation.
-   *
-   * @param admins - The hats admin IDs.
-   * @param details - The hats details fields.
-   * @param maxSupplies - The hats maximum amounts of possible wearers.
-   * @param eligibilityModules - The hats eligibility addresses (zero address is not valid).
-   * @param toggleModules - The hats toggle addresses (zero address is not valid).
-   * @param mutables - True if the hat should be mutable, false otherwise.
-   * @param imageURIs - Optional hats image URIs.
-   * @returns An object containing the call data and the function name.
-   */
-  batchCreateHatsCallData({
-    admins,
-    details,
-    maxSupplies,
-    eligibilityModules,
-    toggleModules,
-    mutables,
-    imageURIs,
-  }: {
-    admins: bigint[];
-    details: string[];
-    maxSupplies: number[];
-    eligibilityModules: Address[];
-    toggleModules: Address[];
-    mutables: boolean[];
-    imageURIs?: string[];
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "batchCreateHats",
-      args: [
-        admins,
-        details,
-        maxSupplies,
-        eligibilityModules,
-        toggleModules,
-        mutables,
-        imageURIs === undefined ? Array(admins.length).fill("") : imageURIs,
-      ],
-    });
-
-    return { functionName: "batchCreateHats", callData };
-  }
-
-  /**
-   * Return the call data for a batchMintHats operation.
-   *
-   * @param hatIds - IDs of the minted hats.
-   * @param wearers - Addresses of the new wearers.
-   * @returns An object containing the call data and the function name.
-   */
-  batchMintHatsCallData({
-    hatIds,
-    wearers,
-  }: {
-    hatIds: bigint[];
-    wearers: Address[];
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "batchMintHats",
-      args: [hatIds, wearers],
-    });
-
-    return { functionName: "batchMintHats", callData };
-  }
-
-  /**
-   * Return the call data for a setHatStatus operation.
-   *
-   * @param hatId - hat ID.
-   * @param newStatus - Hat's new status: true for active, false for inactive.
-   * @returns An object containing the call data and the function name.
-   */
-  setHatStatusCallData({
-    hatId,
-    newStatus,
-  }: {
-    hatId: bigint;
-    newStatus: boolean;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "setHatStatus",
-      args: [hatId, newStatus],
-    });
-
-    return { functionName: "setHatStatus", callData };
-  }
-
-  checkHatStatusCallData({ hatId }: { hatId: bigint }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "checkHatStatus",
-      args: [hatId],
-    });
-
-    return { functionName: "checkHatStatus", callData };
-  }
-
-  /**
-   * Return the call data for a setHatWearerStatus operation.
-   *
-   * @param hatId - Hat ID.
-   * @param wearer - Wearer address.
-   * @param eligible - Wearer's eligibility. True for eligible, false otherwise.
-   * @param standing - Wearer's standing. True for good, false for bad.
-   * @returns An object containing the call data and the function name.
-   */
-  setHatWearerStatusCallData({
-    hatId,
-    wearer,
-    eligible,
-    standing,
-  }: {
-    hatId: bigint;
-    wearer: Address;
-    eligible: boolean;
-    standing: boolean;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "setHatWearerStatus",
-      args: [hatId, wearer, eligible, standing],
-    });
-
-    return { functionName: "setHatWearerStatus", callData };
-  }
-
-  /**
-   * Return the call data for a checkHatWearerStatus operation.
-   *
-   * @param hatId - Hat ID.
-   * @param wearer - Wearer address.
-   * @returns An object containing the call data and the function name.
-   */
-  checkHatWearerStatusCallData({
-    hatId,
-    wearer,
-  }: {
-    hatId: bigint;
-    wearer: Address;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "checkHatWearerStatus",
-      args: [hatId, wearer],
-    });
-
-    return { functionName: "checkHatWearerStatus", callData };
-  }
-
-  /**
-   * Return the call data for a renounceHat operation.
-   *
-   * @param hatId - Hat ID of the hat the caller wishes to renounce.
-   * @returns An object containing the call data and the function name.
-   */
-  renounceHatCallData({ hatId }: { hatId: bigint }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "renounceHat",
-      args: [hatId],
-    });
-
-    return { functionName: "renounceHat", callData };
-  }
-
-  /**
-   * Return the call data for a makeHatImmutable operation.
-   *
-   * @param hatId - Hat ID.
-   * @returns An object containing the call data and the function name.
-   */
-  makeHatImmutableCallData({ hatId }: { hatId: bigint }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "makeHatImmutable",
-      args: [hatId],
-    });
-
-    return { functionName: "makeHatImmutable", callData };
-  }
-
-  /**
-   * Return the call data for a changeHatDetails operation.
-   *
-   * @param hatId - Hat ID.
-   * @param newDetails - The new details.
-   * @returns An object containing the call data and the function name.
-   */
-  changeHatDetailsCallData({
-    hatId,
-    newDetails,
-  }: {
-    hatId: bigint;
-    newDetails: string;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "changeHatDetails",
-      args: [hatId, newDetails],
-    });
-
-    return { functionName: "changeHatDetails", callData };
-  }
-
-  /**
-   * Return the call data for a changeHatEligibility operation.
-   *
-   * @param hatId - Hat ID.
-   * @param newEligibility - The new eligibility address. Zero address is not valid.
-   * @returns An object containing the call data and the function name.
-   */
-  changeHatEligibilityCallData({
-    hatId,
-    newEligibility,
-  }: {
-    hatId: bigint;
-    newEligibility: Address;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "changeHatEligibility",
-      args: [hatId, newEligibility],
-    });
-
-    return { functionName: "changeHatEligibility", callData };
-  }
-
-  /**
-   * Return the call data for a changeHatToggle operation.
-   *
-   * @param hatId - Hat ID.
-   * @param newToggle - The new toggle address. Zero address is not valid.
-   * @returns An object containing the call data and the function name.
-   */
-  changeHatToggleCallData({
-    hatId,
-    newToggle,
-  }: {
-    hatId: bigint;
-    newToggle: Address;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "changeHatToggle",
-      args: [hatId, newToggle],
-    });
-
-    return { functionName: "changeHatToggle", callData };
-  }
-
-  /**
-   * Return the call data for a changeHatImageURI operation.
-   *
-   * @param hatId - Hat ID.
-   * @param newImageURI - The new image URI.
-   * @returns An object containing the call data and the function name.
-   */
-  changeHatImageURICallData({
-    hatId,
-    newImageURI,
-  }: {
-    hatId: bigint;
-    newImageURI: string;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "changeHatImageURI",
-      args: [hatId, newImageURI],
-    });
-
-    return { functionName: "changeHatImageURI", callData };
-  }
-
-  /**
-   * Return the call data for a changeHatMaxSupply operation.
-   *
-   * @param hatId - Hat ID.
-   * @param newMaxSupply -New maximum supply for the hat.
-   * @returns An object containing the call data and the function name.
-   */
-  changeHatMaxSupplyCallData({
-    hatId,
-    newMaxSupply,
-  }: {
-    hatId: bigint;
-    newMaxSupply: number;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "changeHatMaxSupply",
-      args: [hatId, newMaxSupply],
-    });
-
-    return { functionName: "changeHatMaxSupply", callData };
-  }
-
-  /**
-   * Return the call data for a requestLinkTopHatToTree operation.
-   *
-   * @param topHatDomain - The tree domain of the requesting tree. The tree domain is the first four bytes of the tophat ID.
-   * @param requestedAdminHat - ID of the requested new admin hat.
-   * @returns An object containing the call data and the function name.
-   */
-  requestLinkTopHatToTreeCallData({
-    topHatDomain,
-    requestedAdminHat,
-  }: {
-    topHatDomain: number;
-    requestedAdminHat: bigint;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "requestLinkTopHatToTree",
-      args: [topHatDomain, requestedAdminHat],
-    });
-
-    return { functionName: "requestLinkTopHatToTree", callData };
-  }
-
-  /**
-   * Return the call data for a approveLinkTopHatToTree operation.
-   *
-   * @param topHatDomain - The tree domain of the requesting tree. The tree domain is the first four bytes of the tophat ID.
-   * @param newAdminHat - ID of the new admin hat.
-   * @param newEligibility - Optional new eligibility for the linked tophat.
-   * @param newToggle - Optional new toggle for the linked tophat.
-   * @param newDetails - Optional new details for the linked tophat.
-   * @param newImageURI - Optional new image URI for the linked tophat.
-   * @returns An object containing the call data and the function name.
-   */
-  approveLinkTopHatToTreeCallData({
-    topHatDomain,
-    newAdminHat,
-    newEligibility,
-    newToggle,
-    newDetails,
-    newImageURI,
-  }: {
-    topHatDomain: number;
-    newAdminHat: bigint;
-    newEligibility?: Address;
-    newToggle?: Address;
-    newDetails?: string;
-    newImageURI?: string;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "approveLinkTopHatToTree",
-      args: [
-        topHatDomain,
-        newAdminHat,
-        newEligibility === undefined ? ZERO_ADDRESS : newEligibility,
-        newToggle === undefined ? ZERO_ADDRESS : newToggle,
-        newDetails === undefined ? "" : newDetails,
-        newImageURI === undefined ? "" : newImageURI,
-      ],
-    });
-
-    return { functionName: "approveLinkTopHatToTree", callData };
-  }
-
-  /**
-   * Return the call data for a unlinkTopHatFromTree operation.
-   *
-   * @param topHatDomain - The tree domain of the requesting tree. The tree domain is the first four bytes of the tophat ID.
-   * @param wearer - The current wearer of the tophat that is about to be unlinked.
-   * @returns An object containing the call data and the function name.
-   */
-  unlinkTopHatFromTreeCallData({
-    topHatDomain,
-    wearer,
-  }: {
-    topHatDomain: number;
-    wearer: Address;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "unlinkTopHatFromTree",
-      args: [topHatDomain, wearer],
-    });
-
-    return { functionName: "unlinkTopHatFromTree", callData };
-  }
-
-  /**
-   * Return the call data for a relinkTopHatWithinTree operation.
-   *
-   * @param topHatDomain - The tree domain of the requesting tree. The tree domain is the first four bytes of the tophat ID.
-   * @param newAdminHat - ID of the new admin hat.
-   * @param newEligibility - Optional new eligibility for the linked tophat.
-   * @param newToggle - Optional new toggle for the linked tophat.
-   * @param newDetails - Optional new details for the linked tophat.
-   * @param newImageURI - Optional new image URI for the linked tophat.
-   * @returns An object containing the call data and the function name.
-   */
-  relinkTopHatWithinTreeCallData({
-    topHatDomain,
-    newAdminHat,
-    newEligibility,
-    newToggle,
-    newDetails,
-    newImageURI,
-  }: {
-    topHatDomain: number;
-    newAdminHat: bigint;
-    newEligibility?: Address;
-    newToggle?: Address;
-    newDetails?: string;
-    newImageURI?: string;
-  }): {
-    functionName: string;
-    callData: Hex;
-  } {
-    const callData = encodeFunctionData({
-      abi: HATS_ABI,
-      functionName: "relinkTopHatWithinTree",
-      args: [
-        topHatDomain,
-        newAdminHat,
-        newEligibility === undefined ? ZERO_ADDRESS : newEligibility,
-        newToggle === undefined ? ZERO_ADDRESS : newToggle,
-        newDetails === undefined ? "" : newDetails,
-        newImageURI === undefined ? "" : newImageURI,
-      ],
-    });
-
-    return { functionName: "relinkTopHatWithinTree", callData };
-  }
-
-  /**
-   * Get the call data to copy a tree's hats and wearers.
-   * Note: this doensn't include the target's top-hat. The target top-hat should be created separately.
-   *
-   * @param sourceTree - The source tree domain.
-   * @param targetTree - The target tree domain.
-   * @returns An array of call data objects. Passing the result to the multicall function will execute the copy operation.
-   */
-  async copyTreeCallData({
-    sourceTree,
-    targetTree,
-  }: {
-    sourceTree: number;
-    targetTree: number;
-  }): Promise<
-    {
-      functionName: string;
-      callData: Hex;
-    }[]
-  > {
-    if (this._graphqlClient === undefined) {
-      throw new Error("Subgraph client was not initialized");
-    }
-
-    const res: {
-      functionName: string;
-      callData: Hex;
-    }[] = [];
-    const tree = await this._graphqlClient.getTree({
-      chainId: this.chainId,
-      treeId: sourceTree,
-      props: {
-        hats: {
-          details: true,
-          maxSupply: true,
-          imageUri: true,
-          currentSupply: true,
-          levelAtLocalTree: true,
-          eligibility: true,
-          toggle: true,
-          mutable: true,
-          createdAt: true,
-          wearers: {},
-          admin: {},
-        },
-        childOfTree: {},
-        linkedToHat: {},
-        parentOfTrees: {},
-      },
-    });
-
-    const targetTreeHex = treeIdDecimalToHex(targetTree);
-
-    tree.hats?.forEach((hat, index) => {
-      if (index !== 0 && hat.createdAt !== null) {
-        const adminID = hatIdHexToDecimal(
-          targetTreeHex + hat.admin?.id.substring(10)
-        );
-        const createHatCall = this.createHatCallData({
-          admin: adminID,
-          details: hat.details as string,
-          maxSupply: +(hat.maxSupply as string),
-          eligibility: hat.eligibility as `0x${string}`,
-          toggle: hat.toggle as `0x${string}`,
-          mutable: hat.mutable as boolean,
-          imageURI: hat.imageUri as string,
-        });
-        res.push({
-          functionName: "createHat",
-          callData: createHatCall.callData,
-        });
-
-        hat.wearers?.forEach((wearer) => {
-          const mintHatCall = this.mintHatCallData({
-            hatId: hatIdHexToDecimal(targetTreeHex + hat.id.substring(10)),
-            wearer: wearer.id as `0x${string}`,
-          });
-          res.push({ functionName: "mintHat", callData: mintHatCall.callData });
-        });
-      }
-    });
-
-    return res;
   }
 }
